If the status bar is hidden, do not add it to the component.
Previously mouse motion caused the (hidden) status bar to update its label,
which in turn caused revalidation to occur despite the fact that the
label was not visible. To prevent it, the status bar is added and
removed from the component when its visibility is toggled instead of using
JComponent.setVisible.