Fix setting viewer properties before it is shown.
When the size of the image viewer is 0, getImageTransform used to return a non-invertible transformation, which in turn caused an exception for certain operations. The size is now tested before calls to getImageTransform. Also the Javadocs for getImageTransform and pointToPixel has been updated: they now throw IllegalStateException when called on a zero-sized viewer.