Update the status bar on all property changes.
Previously, only setImage calls caused the status bar to update. But zoom changes also cause the pixel under the cursor to change, and there may be other such properties in the future. So let us just ensure that the status bar shows correct information and update on all property changes.