Fix synthetic mouse events when image has overlay.
Due to the presence of the overlay, the code used to think that the mouse exited the component because getMousePosition returned null for the image component. (The overlay basically covered up the image.) Now we use the parent container (which contains the image and all the overlays) to find the mouse cursor.